HDX-10413 indices for data_availability_vat + indices for admin_level
The data_availability_vat table had a primary key that did not include the new "admin_level". This was causing unique constraint violations when trying to import the humanitarian needs table.
I removed the primary key altogether because I don't think it was serving any purpose in the HAPI db and replaced it with a dummy (autoincrementing) one. Also added indices for the other columns that can be used as filters.

Also added indices for each new admin_level column in the other tables. If we switch the HAPI admin_level filter to use these new columns they should have indices.